+++ lib/src/entity/entity_field_base.dart
@@ -58,13 +58,11 @@ abstract class FieldBase<T> {
   void innerLoad(dynamic rawData, {bool copy = false}) {
     if (isComputed) return;
     if (rawData == null) {
-      if (copy) {
-        entity[name] = null;
-      } else {
-        //final defaultValue = this.value;
-        //entity.data.remove(name);
-        entity[name] = innerDefaultValue();
-      }
+      //if (copy) {
+      //  entity[name] = null;
+      //} else {
+      entity[name] = innerDefaultValue();
+      //}
     } else {
       final transformer = _fieldOnLoading ?? ValueTransformers.system();
       entity[name] = transformer(rawData);