+++ lib/src/entity/entity_field_list.dart
@@ -169,8 +169,10 @@ class ListField<E extends Entity> extends FieldBase<List<E>> {
 
   void forEach(void Function(E element) f) => value.forEach(f);
 
-  static ListField<E> create<E extends Entity>() {
-    return Entity().fieldList(E.runtimeType.toString());
+  static ListField<E> create<E extends Entity>({EntityBuilder<E>? register}) {
+    final list = Entity().fieldList<E>(E.runtimeType.toString());
+    if (register != null) list.register(register);
+    return list;
   }
 
   /// Get the ListField (parent) of the entity